C2/requirements.txt
autopep8==2.0.0
click==8.1.3
DateTime==4.7
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
python-decouple==3.6
pytz==2022.6
six==1.16.0
tomli==2.0.1
Werkzeug==2.2.2
zope.interface==5.5.2
autopep8==2.0.0
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
gunicorn==20.1.0
imageio==2.22.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.5
Pillow==9.3.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.16.0
PyJWT==2.6.0
python-decouple==3.6
python-dotenv==0.21.0
six==1.16.0
tomli==2.0.1
Werkzeug==2.2.2
autopep8==2.0.0
cffi==1.15.1
click==8.1.3
cryptography==38.0.4
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
gunicorn==20.1.0
imageio==2.22.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.5
Pillow==9.3.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.16.0
PyJWT==2.6.0
python-decouple==3.6
python-dotenv==0.21.0
six==1.16.0
tomli==2.0.1
Werkzeug==2.2.2
